refactor(codex): split MCP attestation helper
